refactor(#1374): rename route-error -> RouteRetryError for PascalCase consistency (review nit)
