+++ packages/web/src/routes/$locale/_admin/admin/revenue.tsx
@@ -1,9 +1,9 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { AnomaliesPanel } from '@/vite/admin/AnomaliesPanel'
 import { RevenueView } from '@/vite/admin/RevenueView'
 import { paymentAnomaliesQueryOptions } from '@/vite/admin/anomalies/api'
 import { adminRevenueQueryOptions } from '@/vite/admin/revenue/api'
-import { RouteRetryError } from '@/vite/route-error'
 import { useQuery, useSuspenseQuery } from '@tanstack/react-query'
 import { type ErrorComponentProps, createFileRoute } from '@tanstack/react-router'
 import { useTranslations } from 'use-intl'
