+++ packages/web/src/routes/$locale/_business/dashboard.tsx
@@ -1,9 +1,9 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { useOperatorContext } from '@/vite/operator-context'
 import { OperatorDashboardView } from '@/vite/operator-dashboard/OperatorDashboardView'
 import { operatorOverviewQueryOptions } from '@/vite/operator-dashboard/api'
 import { operatorFleetQueryOptions } from '@/vite/operator-fleet/api'
-import { RouteRetryError } from '@/vite/route-error'
 import { useSession } from '@/vite/session'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { type ErrorComponentProps, createFileRoute } from '@tanstack/react-router'
