+++ packages/web/src/routes/$locale/_business/manage/add-ons.tsx
@@ -1,9 +1,9 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { DEFAULT_LOCALE, isLocale } from '@/vite/i18n/locale'
 import { OperatorAddOnsView } from '@/vite/operator-add-ons/OperatorAddOnsView'
 import { addOnsQueryOptions } from '@/vite/operator-add-ons/api'
 import { useOperatorScope } from '@/vite/operator-context'
-import { RouteRetryError } from '@/vite/route-error'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { type ErrorComponentProps, createFileRoute } from '@tanstack/react-router'
 import { useLocale, useTranslations } from 'use-intl'
