+++ packages/web/src/routes/$locale/_business/manage/bookings/$bookingId.tsx
@@ -1,4 +1,5 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { isOperatorSession } from '@/vite/guards'
 import { BookingActionsPanel } from '@/vite/operator-bookings/BookingActionsPanel'
 import { BookingTimeline } from '@/vite/operator-bookings/BookingTimeline'
@@ -10,7 +11,6 @@ import {
   substitutionCandidatesQueryOptions,
 } from '@/vite/operator-bookings/api'
 import { RateRenterPanel } from '@/vite/reviews'
-import { RouteRetryError } from '@/vite/route-error'
 import { sessionQueryOptions } from '@/vite/session'
 import { useQuery, useSuspenseQuery } from '@tanstack/react-query'
 import { type ErrorComponentProps, Link, createFileRoute, notFound } from '@tanstack/react-router'
