+++ packages/web/src/routes/$locale/_business/manage/fees.tsx
@@ -1,10 +1,10 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { isOperatorSession } from '@/vite/guards'
 import { operatorClassesQueryOptions } from '@/vite/operator-classes/api'
 import { useOperatorScope } from '@/vite/operator-context'
 import { OperatorFeesView } from '@/vite/operator-fees/OperatorFeesView'
 import { feeSchedulesQueryOptions } from '@/vite/operator-fees/api'
-import { RouteRetryError } from '@/vite/route-error'
 import { useSession } from '@/vite/session'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { type ErrorComponentProps, createFileRoute } from '@tanstack/react-router'
