+++ packages/web/src/routes/$locale/_business/manage/fleet/$vehicleId.tsx
@@ -1,9 +1,9 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { canWriteAsOperator } from '@/vite/guards'
 import { useOperatorContext } from '@/vite/operator-context'
 import { VehicleDetail } from '@/vite/operator-fleet/VehicleDetail'
 import { vehicleDetailQueryOptions } from '@/vite/operator-fleet/api'
-import { RouteRetryError } from '@/vite/route-error'
 import { sessionQueryOptions } from '@/vite/session'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { type ErrorComponentProps, Link, createFileRoute, notFound } from '@tanstack/react-router'
