+++ packages/web/src/routes/$locale/_business/manage/insurance.tsx
@@ -1,8 +1,8 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { useOperatorScope } from '@/vite/operator-context'
 import { OperatorInsuranceView } from '@/vite/operator-insurance/OperatorInsuranceView'
 import { insuranceOptionsQueryOptions } from '@/vite/operator-insurance/api'
-import { RouteRetryError } from '@/vite/route-error'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { type ErrorComponentProps, createFileRoute } from '@tanstack/react-router'
 import { useTranslations } from 'use-intl'
