+++ packages/web/src/routes/$locale/_business/manage/locations.tsx
@@ -1,8 +1,8 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { useOperatorScope } from '@/vite/operator-context'
 import { OperatorLocationsView } from '@/vite/operator-locations/OperatorLocationsView'
 import { operatorLocationsQueryOptions } from '@/vite/operator-locations/api'
-import { RouteRetryError } from '@/vite/route-error'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { type ErrorComponentProps, createFileRoute } from '@tanstack/react-router'
 import { useTranslations } from 'use-intl'
