+++ packages/web/src/routes/$locale/_business/manage/messages/index.tsx
@@ -1,6 +1,6 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { OperatorInboxView, threadsQueryOptions } from '@/vite/messaging'
-import { RouteRetryError } from '@/vite/route-error'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { type ErrorComponentProps, createFileRoute } from '@tanstack/react-router'
 import { useTranslations } from 'use-intl'
