+++ packages/web/src/routes/$locale/_business/manage/team.tsx
@@ -1,5 +1,6 @@
 import { Button } from '@/components/ui/button'
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { featureFlagsQueryOptions, resolveFeatureFlag } from '@/vite/config'
 import { isOperatorOwnerSession, isOperatorSession } from '@/vite/guards'
 import { DeactivateMemberDialog } from '@/vite/operator-team/DeactivateMemberDialog'
@@ -7,7 +8,6 @@ import { InviteStaffDialog } from '@/vite/operator-team/InviteStaffDialog'
 import { RevokeInviteDialog } from '@/vite/operator-team/RevokeInviteDialog'
 import { TeamView } from '@/vite/operator-team/TeamView'
 import { teamInvitesQueryOptions, teamMembersQueryOptions } from '@/vite/operator-team/api'
-import { RouteRetryError } from '@/vite/route-error'
 import { sessionQueryOptions } from '@/vite/session'
 import type { OperatorInviteData, OperatorMemberData } from '@kuruma/shared/types/operator-team'
 import { useSuspenseQuery } from '@tanstack/react-query'
