+++ packages/web/src/routes/$locale/_renter/bookings/index.tsx
@@ -1,9 +1,9 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { MyBookingsView } from '@/vite/bookings/MyBookingsView'
 import { myBookingsQueryOptions } from '@/vite/bookings/api'
 import { indexThreadIdsByBooking, threadsQueryOptions } from '@/vite/messaging'
 import { renterReviewedSubjects, reviewsForBookingQueryOptions } from '@/vite/reviews'
-import { RouteRetryError } from '@/vite/route-error'
 import { sessionQueryOptions } from '@/vite/session'
 import { useQueries, useQuery, useSuspenseQuery } from '@tanstack/react-query'
 import { type ErrorComponentProps, createFileRoute, redirect } from '@tanstack/react-router'
