+++ packages/web/src/routes/$locale/_renter/messages/$threadId.tsx
@@ -1,7 +1,7 @@
 import { PageSkeleton } from '@/vite/PageSkeleton'
+import { RouteRetryError } from '@/vite/RouteRetryError'
 import { bookingByIdQueryOptions } from '@/vite/bookings/api'
 import { ConversationView, threadByIdQueryOptions, usersByIdsQueryOptions } from '@/vite/messaging'
-import { RouteRetryError } from '@/vite/route-error'
 import { sessionQueryOptions } from '@/vite/session'
 import { useQuery, useSuspenseQuery } from '@tanstack/react-query'
 import {
